requirements.txt
streamlit==1.42.0
google-generativeai==0.8.4
PyPDF2==3.0.1
pandas==2.2.3
